rag_system/requirements.txt
# RAG System Requirements
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pypdf>=3.15.0
python-docx>=0.8.11
pandas>=2.0.0
numpy>=1.20.0
streamlit>=1.25.0
openai>=1.0.0
faiss-cpu>=1.7.0
tiktoken>=0.5.0 